+++ examples/sampleplayertimestretched/src/SampleplayertimestretchedProcessor.h
@@ -310,7 +310,7 @@ class AudioPluginAudioProcessor : public juce::AudioProcessor, public juce::Audi
             {"type",
              [](AudioPluginAudioProcessor& p, const float v)
              {
-                 p.pluginRunner->setType(static_cast<int>(v));
+                 p.pluginRunner->setType(static_cast<size_t>(v));
                  p.m_fileIo.updateParameter(PatchParameters::Id::type, v);
              }},
             {"position",
